+++ include/k3/tok3n/parsers/compound/Sequence.h
@@ -105,7 +105,7 @@ struct Sequence
 		using Executor = detail::SequenceExecutor<void, V>;
 		Executor executor{ .input = input };
 
-		bool successful = (... && executor.execute<Ps, -1, _trait<V>::unwrapped>());
+		bool successful = (... && executor.execute<Ps, static_cast<std::size_t>(-1), _trait<V>::unwrapped>());
 
 		if (successful)
 			return Result<void, V>{ success, executor.input };
